q-8 and q-9 updated
